Yaml file with six updated maps
Here are the updates for the yaml file. I hope you can still see it in comparison with the old file and spot all the changes. It should only be my six maps that that have changed. I tested it out locally, downloaded all the maps and tested out the maps.

As I mentioned elsewhere, there is a background error reported in the console when running maps. But I can’t see it being related to my maps specifically. Downloading and running any other map results in the exact same error, even without using my updated yaml file.

We are talking about the "duplicate map” / ”Found too many files for…” error discussed here:
https://github.com/triplea-game/triplea/issues/855
https://github.com/triplea-game/triplea/pull/620